refactor ModalAddEditCard with custom hook. Works. Need refactor on rtk state
